0.965 ナビの数値の丸めを直し、航跡描画の無駄を削った
■ 旋回角速度に ±0.67deg/s の量子化ノイズが乗っていた
update_degpersecond() が真方位を int で受けていた。窓の中で差分を積むと
切り捨て誤差は両端の 2 つだけが残るため、誤差 1 度 ÷ 窓 1.5 秒 = ±0.67deg/s。
真値 1.00deg/s が開始位相によって 0.67 にも 1.33 にもなる計算で、
着色 1.0 / 強調 3.0 / コース警報の「修正中」判定 0.5deg/s に対して無視できない。
配列は元から float、GNSS も NAV-PVT の headMotion（分解能 1e-5 度）なので
捨てていたのはここだけだった。float で受けるようにした。

■ truec の切り捨て
(int) キャストで常に最大 1 度小さい側へ寄り、そのズレが steer_angle
（赤い修正矢印としきい値判定）にも乗っていた。float で保持し、
表示のときだけ四捨五入する。正規化は fmodf で全域を確認した。

■ 修正矢印のしきい値が実質 1 度ずれていた
draw_triangle() が steer_angle を int で受けていたため、15/55/100 度が
実質 16/56/101 度になっていた。float で受ける。
draw_course_warning() も同様（|ズレ|<1 度 で左右の表示が逆になり得た）。

■ 航跡描画が座標変換を 4 倍やっていた
・draw_track() が getData(i) と getData(i+1) をそれぞれ変換していたので、
  各点が 2 回ずつ変換されていた。前回の結果を持ち回るようにした。
・latLonToXY() が呼ばれるたびに地図中心の Mercator Y（log(tan())）を
  計算し直していた。1 フレームの間ずっと同じ値なので、入力が変わったときだけ
  計算するようにした。航跡 500 点で log+tan が約 2000 回 → 約 500 回になる。

■ その他
・calculatePointD() が A==B のとき出力を書かずに return していた。
  呼び出し側は未初期化のローカルを渡しているので、この経路に入ると
  誘導線が明後日を指す。始点をそのまま返すようにした（現状は到達しない）。
・link_get_truetrack() だけ範囲チェックが無く、CRC を通った壊れフレームで
  655 度までの値がナビ計算に入り得た。実 GPS / リプレイ / デモと同じ形に揃えた。
・cord_tft::isOutsideTft() が SCREEN_HEIGHT(320) で判定していたが、
  cord_tft は backscreen(240x240) 内の座標なので BACKSCREEN_SIZE に直した。
・使われていない cord_tft points[MAX_TRACK_CORDS] を削除
  （リンカが既に捨てていたので RAM は変わらない）。
・1km 円の緯度補正も 35.29 に揃えた（10.975km 円だけ直っていた）。

RELEASE と非 RELEASE の両構成でビルド確認済み。新規の警告なし。

+++ GPS_TFT_map.ino
@@ -7,7 +7,7 @@
 //           地図背景はフラッシュ内蔵のベクタ地図（vectormap.cpp）を使う。
 //           SDカード上のBMPタイル方式は廃止済み。
 // Author  : MasaoC (@masao_mobile)
-// Updated : 2026/09/13
+// Updated : 2026/09/14
 // ============================================================
 //
 // ■ src/ に置いてあるもの
@@ -116,7 +116,7 @@ extern volatile bool loading_sddetail;
 extern bool sd_detail_loading_displayed;
 
 void reset_degpersecond();
-void update_degpersecond(int true_track);
+void update_degpersecond(float true_track);
 void check_destination_toofar();
 static void apply_auto10k_status(int st);
 void update_course_warning(float degpersecond);
@@ -1502,7 +1502,7 @@ void reset_degpersecond() {
 //   4. 配列を1つずらして古いサンプルを捨てる
 //   ※ 360度またぎ（例: 359→1度）を -180〜+180 に正規化して計算する
 //   ※ 時間正規化により GPS レート（1Hz / 2Hz）に依存しない
-void update_degpersecond(int true_track) {
+void update_degpersecond(float true_track) {
   uint32_t now = millis();
   truetrack_samples[sampleIndex] = true_track;
   truetrack_sample_times[sampleIndex] = now;
@@ -1612,7 +1612,13 @@ static void nav_alarm_tick(bool new_gps_info) {
 
   // GPS コース更新時（1Hz または 2Hz）にのみ実行する処理
   if (newcourse_arrived) {
-    int ttrack = get_gps_truetrack();
+    // ★ float で受けること。以前は int で、切り捨て誤差が窓の両端に残り
+    //   deg/s に最大 ±0.67（誤差1度 ÷ 窓1.5秒）の量子化ノイズが乗っていた。
+    //   着色 1.0 / 強調 3.0 / コース警報の「修正中」判定 0.5 deg/s という
+    //   しきい値に対して無視できない大きさだった。
+    //   truetrack_samples[] は元から float、GNSS 側も NAV-PVT の headMotion
+    //   （分解能 1e-5 度）なので、捨てていたのはここだけ。
+    float ttrack = get_gps_truetrack();
 
     // ★ **この順番を入れ替えないこと。**
     //   nav_update() → steer_angle → update_course_warning() の順でなければ、
@@ -1624,6 +1630,7 @@ static void nav_alarm_tick(bool new_gps_info) {
 
     // 針路誤差 (steer_angle) の計算:
     // truec はナビが指示する真方位コース、ttrack は GPS 実測の真方位。
+    // どちらも小数を保った値なので、結果もそのまま使える（表示だけ丸める）。
     // 結果を -180〜+180 に正規化する。
     // ※ 以前は (magc - 8) と書いていたが、magc は真方位に +8 した値だったので
     //   -8 はそれを打ち消していただけ。つまり計算は元から真方位同士で、
